test: ping-pong End Point 테스트코드 작성
- 앱실행환경을 .env파일로 관리
- babel모듈 3가지 추가 및 .babelrc 팡딜작성
- models/index.js에서 개발환경 변수를 .env로에서 관리
- app.js 에서 실행환경이 test 일시에 DB connection을 실행하지 않음
- jest supertest 모듈 추가

+++ app.js
@@ -7,13 +7,18 @@ const { sequelize } = require("./models");
 const ccqp = require("ccqp");
 const swaggerUi = require("swagger-ui-express");
 const swaggerDocument = require("./swagger/swagger-output.json");
+const { NODE_ENV } = require("./config");
 
 const app = express();
 
-sequelize
-  .sync({ force: false })
-  .then(() => console.log("connected database"))
-  .catch((err) => console.error("occurred error in database connecting", err));
+if (NODE_ENV !== "test") {
+  sequelize
+    .sync({ force: false })
+    .then(() => console.log("connected database"))
+    .catch((err) =>
+      console.error("occurred error in database connecting", err)
+    );
+}
 
 app.use(cors());
 app.use(logger("combined"));
